+++ terraform/outputs.tf
@@ -1,7 +1,0 @@
-output "registry_latest" {
-  value = "${data.docker_registry_image.latest.sha256_digest}"
-}
-
-output "new_image" {
-  value = "${docker_image.new.latest}"
-}