+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ADD Terraform scripts for infrastructural automation
 - ADD statistics capturing and `!stats` command for admins to view
+- ADD `!flight` series of commands for UOAF member to manage pickup flights - similar system to `!lfg` commands
 - UPDATE `!ready` to inform when an alarm was overridden
 
 ## v2.5.0